+++ RELEASES.md
@@ -6,7 +6,9 @@ Ladybug Test Tool release notes
 Upcoming
 --------
 
-...
+- Replace Download all with Download table and Download tree
+- Add Open all
+- Move error label from above to below buttons in debug tab